refactor: restore HomePage to default content
