+++ src/pages/home/HomePage.tsx
@@ -1,0 +1,13 @@
+import * as S from './HomePage.styles';
+
+/**
+ * Home Page Component
+ * TODO: 실제 홈페이지 내용으로 교체
+ */
+export default function HomePage() {
+  return (
+    <S.Container>
+
+    </S.Container>
+  );
+}
